refactor : 자동차 이름 나누기  기능을 Cars객체로 옮기기
