+++ src/main/java/racingcar/Application.java
@@ -1,7 +1,9 @@
 package racingcar;
 
+import racingcar.controller.Controller;
+
 public class Application {
     public static void main(String[] args) {
-        // TODO: 프로그램 구현
+        Controller controller = new Controller();
     }
 }
