+++ src/main/java/racingcar/model/Move.java
@@ -1,5 +1,6 @@
 package racingcar.model;
 
+import java.util.List;
 import java.util.Random;
 
 public class Move {
