+++ src/main/java/racingcar/view/outputView.java
@@ -1,0 +1,5 @@
+package racingcar.view;
+
+public class outputView {
+
+}
